requirements.dev.txt
pytest-env==0.6.2
coverage==4.5.4
factory-boy==2.12.0
Faker==1.0.8
freezegun==0.3.12
pytest==4.4.0
pytest-cov==2.6.1
pytest-mock==1.10.3
pytest-sugar==0.9.2
responses==0.10.6
testfixtures==6.10.0
black==19.3b0
pytest-watch
pytest-flask
